int userIDをlongに変更

+++ qa_app/src/main/java/com/example/qa_app/repository/SetupRepository.java
@@ -10,17 +10,17 @@ public class SetupRepository {
     
     private final JdbcTemplate jdbc;
 
-    public void incomeInsert(int userId, boolean financialCategory, Integer monthlyAmount) {
+    public void incomeInsert(long userId, boolean financialCategory, Integer monthlyAmount) {
         String sql = "INSERT INTO daily(user_id, financial_category, monthly_amount)VALUES (?, ?, ?, ?)";
         jdbc.update(sql, userId, financialCategory, monthlyAmount);
     }
 
-    public void fixedInsert(int userId, boolean financialCategory, String fixedName, Integer monthlyAmount) {
+    public void fixedInsert(long userId, boolean financialCategory, String fixedName, Integer monthlyAmount) {
         String sql = "INSERT INTO daily(user_id, financial_category, fixed_name, monthly_amount)VALUES (?, ?, ?, ?)";
         jdbc.update(sql, userId, financialCategory, fixedName, monthlyAmount);
     }
 
-    public void categoriesInsert(int userId, String categoryName) {
+    public void categoriesInsert(long userId, String categoryName) {
         String sql = "INSERT INTO categories(user_id,category_name)VALUES (?, ?)";
         jdbc.update(sql, userId, categoryName);
     }
